requirements.txt
alembic==1.8.1
asgiref==3.5.2
bcrypt==4.0.1
click==8.1.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
python-decouple==3.6
python-dotenv==0.21.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
Werkzeug==2.2.2
WTForms==3.0.1
